api/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.6
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
